safetoswim/servers/requirements.txt
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.6.2
anaconda-project==0.8.2
asn1crypto==0.24.0
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
awsebcli==3.12.4
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
botocore==1.9.6
Bottleneck==1.2.1
brewer2mpl==1.4.1
buildbot==1.0.0
bz2file==0.98
cement==2.8.2
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
cntk==2.4
colorama==0.3.7
comtypes==1.1.4
conda==4.4.7
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
dill==0.2.7.1
distributed==1.20.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
gensim==3.0.1
gevent==1.2.2
ggplot==0.6.5
glob2==0.6
greenlet==0.4.12
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
Keras==2.1.2
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.0
numpydoc==0.7.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
pathspec==0.5.5
patsy==0.5.0
pbr==3.1.1
pep8==1.7.1
pew==1.1.2
pickleshare==0.7.4
Pillow==5.0.0
pipenv==11.1.4
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==5.4.3
py==1.5.2
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.0
pylint==1.8.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyreadline==2.1
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
pywin32==222
pywinpty==0.5
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
-e git+https://github.com/TheTimKiely/RembrandtML@9a7915f476557f19cc5990f3982239dfda667301#egg=rembrandtml
requests==2.9.1
requests-toolbelt==0.8.0
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
semantic-version==2.5.0
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.7.5
tblib==1.3.2
Tempita==0.5.2
tensorflow==1.3.0
tensorflow-tensorboard==0.1.6
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.1+19.gf62fd3942
toolz==0.9.0
torch==0.2.1+a4fc05a
tornado==4.5.3
tqdm==4.19.6
traitlets==4.3.2
twine==1.10.0
Twisted==17.9.0
txaio==2.9.0
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.47.0
Werkzeug==0.14.1
widgetsnbextension==3.1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xgboost==0.7
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
zict==0.1.3
zope.interface==4.4.3
